refact model
