+++ deepmd/model/__init__.py
@@ -1,0 +1,5 @@
+from .ener import EnerModel
+from .tensor import WFCModel
+from .tensor import DipoleModel
+from .tensor import PolarModel
+from .tensor import GlobalPolarModel
