+++ deepmd/model/ener.py
@@ -1,92 +1,65 @@
 import numpy as np
+from typing import Tuple, List
+
 from deepmd.env import tf
-from collections import defaultdict
 from deepmd.utils.pair_tab import PairTab
 from deepmd.common import ClassArg
-
 from deepmd.run_options import global_cvt_2_ener_float, MODEL_VERSION
 from deepmd.env import op_module
+from .model_stat import make_stat_input, merge_sys_stat
 
-
-def _make_all_stat_ref(data, nbatches):
-    all_stat = defaultdict(list)
-    for ii in range(data.get_nsystems()) :
-        for jj in range(nbatches) :
-            stat_data = data.get_batch (sys_idx = ii)
-            for dd in stat_data:
-                if dd == "natoms_vec":
-                    stat_data[dd] = stat_data[dd].astype(np.int32) 
-                all_stat[dd].append(stat_data[dd])        
-    return all_stat
-
-
-def make_stat_input(data, nbatches, merge_sys = True):
-    """
-    pack data for statistics
-    Parameters
-    ----------
-    data:
-        The data
-    merge_sys: bool (True)
-        Merge system data
-    Returns
-    -------
-    all_stat:
-        A dictionary of list of list storing data for stat. 
-        if merge_sys == False data can be accessed by 
-            all_stat[key][sys_idx][batch_idx][frame_idx]
-        else merge_sys == True can be accessed by 
-            all_stat[key][batch_idx][frame_idx]
-    """
-    all_stat = defaultdict(list)
-    for ii in range(data.get_nsystems()) :
-        sys_stat =  defaultdict(list)
-        for jj in range(nbatches) :
-            stat_data = data.get_batch (sys_idx = ii)
-            for dd in stat_data:
-                if dd == "natoms_vec":
-                    stat_data[dd] = stat_data[dd].astype(np.int32) 
-                sys_stat[dd].append(stat_data[dd])
-        for dd in sys_stat:
-            if merge_sys:
-                for bb in sys_stat[dd]:
-                    all_stat[dd].append(bb)
-            else:                    
-                all_stat[dd].append(sys_stat[dd])
-    return all_stat
-
-def merge_sys_stat(all_stat):
-    first_key = list(all_stat.keys())[0]
-    nsys = len(all_stat[first_key])
-    ret = defaultdict(list)
-    for ii in range(nsys):
-        for dd in all_stat:
-            for bb in all_stat[dd][ii]:
-                ret[dd].append(bb)
-    return ret
-
-
-class Model() :
+class EnerModel() :
     model_type = 'ener'
 
-    def __init__ (self, jdata, descrpt, fitting):
+    def __init__ (
+            self, 
+            descrpt, 
+            fitting, 
+            type_map : List[str],
+            data_stat_nbatch : int = 10,
+            data_stat_protect : float = 1e-2,
+            use_srtab : str = None,
+            smin_alpha : float = None,
+            sw_rmin : float = None,
+            sw_rmax : float = None
+    ) -> None:
+        """
+        Constructor
+
+        Parameters
+        ----------
+        descrpt
+                Descriptor
+        fitting
+                Fitting net
+        type_map
+                Mapping atom type to the name (str) of the type.
+                For example `type_map[1]` gives the name of the type 1.
+        data_stat_nbatch
+                Number of frames used for data statistic
+        data_stat_protect
+                Protect parameter for atomic energy regression
+        use_srtab
+                The table for the short-range pairwise interaction added on top of DP. The table is a text data file with (N_t + 1) * N_t / 2 + 1 columes. The first colume is the distance between atoms. The second to the last columes are energies for pairs of certain types. For example we have two atom types, 0 and 1. The columes from 2nd to 4th are for 0-0, 0-1 and 1-1 correspondingly.
+        smin_alpha
+                The short-range tabulated interaction will be swithed according to the distance of the nearest neighbor. This distance is calculated by softmin. This parameter is the decaying parameter in the softmin. It is only required when `use_srtab` is provided.
+        sw_rmin
+                The lower boundary of the interpolation between short-range tabulated interaction and DP. It is only required when `use_srtab` is provided.
+        sw_rmin
+                The upper boundary of the interpolation between short-range tabulated interaction and DP. It is only required when `use_srtab` is provided.
+        """
+        # descriptor
         self.descrpt = descrpt
         self.rcut = self.descrpt.get_rcut()
         self.ntypes = self.descrpt.get_ntypes()
         # fitting
         self.fitting = fitting
         self.numb_fparam = self.fitting.get_numb_fparam()
-
-        args = ClassArg()\
-               .add('type_map',         list,   default = []) \
-               .add('data_stat_nbatch', int,    default = 10) \
-               .add('data_stat_protect',float,  default = 1e-2) \
-               .add('use_srtab',        str)
-        class_data = args.parse(jdata)
-        self.type_map = class_data['type_map']
-        self.srtab_name = class_data['use_srtab']
-        self.data_stat_nbatch = class_data['data_stat_nbatch']
-        self.data_stat_protect = class_data['data_stat_protect']
+        # other inputs
+        self.type_map = type_map
+        self.data_stat_nbatch = data_stat_nbatch
+        self.data_stat_protect = data_stat_protect
+        self.srtab_name = use_srtab
         if self.srtab_name is not None :
             self.srtab = PairTab(self.srtab_name)
             args.add('smin_alpha',      float,  must = True)\
@@ -264,127 +237,4 @@ class Model() :
         model_dict['atype'] = atype
         
         return model_dict
-
-
-class TensorModel() :
-    def __init__ (self, jdata, descrpt, fitting, var_name):
-        self.model_type = var_name        
-        self.descrpt = descrpt
-        self.rcut = self.descrpt.get_rcut()
-        self.ntypes = self.descrpt.get_ntypes()
-        # fitting
-        self.fitting = fitting
-
-        args = ClassArg()\
-               .add('type_map',         list,   default = []) \
-               .add('data_stat_nbatch', int,    default = 10) \
-               .add('data_stat_protect',float,  default = 1e-2)
-        class_data = args.parse(jdata)
-        self.type_map = class_data['type_map']
-        self.data_stat_nbatch = class_data['data_stat_nbatch']
-        self.data_stat_protect = class_data['data_stat_protect']
-    
-    def get_rcut (self) :
-        return self.rcut
-
-    def get_ntypes (self) :
-        return self.ntypes
-
-    def get_type_map (self) :
-        return self.type_map
-
-    def get_sel_type(self):
-        return self.fitting.get_sel_type()
-
-    def get_out_size (self) :
-        return self.fitting.get_out_size()
-
-    def data_stat(self, data):
-        all_stat = make_stat_input(data, self.data_stat_nbatch, merge_sys = False)
-        m_all_stat = merge_sys_stat(all_stat)        
-        self._compute_input_stat (m_all_stat, protection = self.data_stat_protect)
-        self._compute_output_stat(all_stat)
-
-    def _compute_input_stat(self, all_stat, protection = 1e-2) :
-        self.descrpt.compute_input_stats(all_stat['coord'],
-                                         all_stat['box'],
-                                         all_stat['type'],
-                                         all_stat['natoms_vec'],
-                                         all_stat['default_mesh'], 
-                                         all_stat)
-        if hasattr(self.fitting, 'compute_input_stats'):
-            self.fitting.compute_input_stats(all_stat, protection = protection)
-
-    def _compute_output_stat (self, all_stat) :
-        if hasattr(self.fitting, 'compute_output_stats'):
-            self.fitting.compute_output_stats(all_stat)
-
-    def build (self, 
-               coord_, 
-               atype_,
-               natoms,
-               box, 
-               mesh,
-               input_dict,
-               suffix = '', 
-               reuse = None):
-        with tf.variable_scope('model_attr' + suffix, reuse = reuse) :
-            t_tmap = tf.constant(' '.join(self.type_map), 
-                                 name = 'tmap', 
-                                 dtype = tf.string)
-            t_st = tf.constant(self.get_sel_type(), 
-                               name = 'sel_type',
-                               dtype = tf.int32)
-            t_mt = tf.constant(self.model_type, 
-                               name = 'model_type', 
-                               dtype = tf.string)
-            t_ver = tf.constant(MODEL_VERSION,
-                                name = 'model_version',
-                                dtype = tf.string)
-            t_od = tf.constant(self.get_out_size(), 
-                               name = 'output_dim', 
-                               dtype = tf.int32)
-
-        dout \
-            = self.descrpt.build(coord_,
-                                 atype_,
-                                 natoms,
-                                 box,
-                                 mesh,
-                                 input_dict,
-                                 suffix = suffix,
-                                 reuse = reuse)
-        dout = tf.identity(dout, name='o_descriptor')
-        rot_mat = self.descrpt.get_rot_mat()
-        rot_mat = tf.identity(rot_mat, name = 'o_rot_mat')
-
-        output = self.fitting.build (dout, 
-                                     rot_mat,
-                                     natoms, 
-                                     reuse = reuse, 
-                                     suffix = suffix)
-        output = tf.identity(output, name = 'o_' + self.model_type)
-
-        return {self.model_type: output}
-
-
-class WFCModel(TensorModel):
-    def __init__(self, jdata, descrpt, fitting) :
-        TensorModel.__init__(self, jdata, descrpt, fitting, 'wfc')
-
-
-class DipoleModel(TensorModel):
-    def __init__(self, jdata, descrpt, fitting) :
-        TensorModel.__init__(self, jdata, descrpt, fitting, 'dipole')
-
-
-class PolarModel(TensorModel):
-    def __init__(self, jdata, descrpt, fitting) :
-        TensorModel.__init__(self, jdata, descrpt, fitting, 'polar')
-
-
-class GlobalPolarModel(TensorModel):
-    def __init__(self, jdata, descrpt, fitting) :
-        TensorModel.__init__(self, jdata, descrpt, fitting, 'global_polar')
-
