+++ source/train/CMakeLists.txt
@@ -2,7 +2,7 @@
 
 configure_file("run_config.ini" "${CMAKE_CURRENT_BINARY_DIR}/run_config.ini" @ONLY)
 
-file(GLOB LIB_PY main.py calculator.py model.py trainer.py run_options.py )
+file(GLOB LIB_PY main.py calculator.py trainer.py run_options.py )
 
 install(
   FILES		${LIB_PY}
